Update CD2LayoutConverter.java
Reorganized the code so that identical sequences are now in separate methods, hence making it easier to maintain the code. The center point of reactions is now correctly used and the reaction glyph is defined as a small square box whose center is the center of the reaction now. Before the reaction glyph was defined as a large box around the entire reaction. Also fixed a number of smaller bugs.