Improved CD2LayoutConverter and did minor changes elsewhere.
* Set log level to fine for message in SBMLUtil when removing empty
listOf objects.
* Minor change in LayoutUtil for improved code readability
* Multiple corrections and unifications in CD2LayoutConverter:
a) Unified the creation of TextGlyphs by refactoring and putting this
into a separate method.
b) Corrected error that wrong dimensions were used for setting the
depth, which resulted in unspecified depth for text glyphs.
c) Refactoring to shift the creation of compartmentGlyphs into one
method.
d) Created a new method to convert just one ReactionWrapper to a
ReactionGlyph instead of dealing with the entire list at once.
e) Improved the calculation of bounding boxes for speciesReferenceGlyphs
to include the full arc.
f) Improved error and log messages
g) Corrected error when creating speciesReferenceGlyphs: the link to the
core speciesReference was left blank because these were lacking ids.
Note that ids on speciesReference were only introduced in L3V1, so the
CellDesigner model lacks them. Ids are now created when first needed.
h) The calculation of the size of bounding boxes for textGlyphs was
improved and made more realistic, tested on several examples.